feat: player movement, options to select, refactor screen and size init

+++ src/score_handler.rs
@@ -4,7 +4,7 @@ use std::{
     io::{BufRead, BufReader},
 };
 
-pub fn save_high_score(score: u16) -> std::io::Result<()> {
+pub fn save_high_score(score: u64) -> std::io::Result<()> {
     let mut file = OpenOptions::new()
         .write(true)
         .create(true)
@@ -13,12 +13,12 @@ pub fn save_high_score(score: u16) -> std::io::Result<()> {
     Ok(())
 }
 
-pub fn load_high_score() -> u16 {
+pub fn load_high_score() -> u64 {
     if let Ok(file) = File::open("high_score.txt") {
         let reader = BufReader::new(file);
         if let Some(Ok(line)) = reader.lines().next() {
             if let Ok(score) = line.parse::<u16>() {
-                return score;
+                return score as u64;
             }
         }
     }
